// web/src/enums/dictTypeCode.ts
export enum KEnum {
  BIZ_TYPE = 'K_BIZ_TYPE',
  PRODIVER = 'K_PRODIVER',
  SCHEDULE_TYPE = 'K_SCHEDULE_TYPE',
  IS_COVER_DIFF_ISP = 'K_IS_COVER_DIFF_ISP',
  DELIVERY_STATUS = 'K_DELIVERY_STATUS',
  DELIVERY_TYPE = 'K_DELIVERY_TYPE',
  DEVICE_STATUS = 'K_DEVICE_STATUS',
  NETWORK_TYPE = 'K_NETWORK_TYPE',
  ISP_NAME = 'K_ISP_NAME',
  DEVICE_TYPE = 'K_DEVICE_TYPE',
  SAME_AREA_REPLACE = 'K_SAME_AREA_REPLACE',
  HDD_TRANSFORM = 'K_HDD_TRANSFORM',
  ISP = 'K_ISP',
  OP_STATUS = 'K_OP_STATUS',
  IS_DISABLED = 'K_IS_DISABLED',
  BUSINESS_STATUS = 'K_BUSINESS_STATUS',
  DEVICE_CATEGORY = 'K_DEVICE_CATEGORY',
  PROVIDER_TYPE = 'K_PROVIDER_TYPE',
  OCCUPY_STATUS = 'K_OCCUPY_STATUS',
  HDD_LOSE_REASON = 'K_HDD_LOSE_REASON',
  PARTITION_MOUNT_STATUS = 'K_PARTITION_MOUNT_STATUS',
}

export enum SysEnum {
  USER_FROM = 'SYS_USER_FROM',
  REQUEST_METHOD = 'SYS_REQUEST_METHOD',
  HANDLE_SOURCE = 'SYS_HANDLE_SOURCE',
}

export enum CjEnum {
  CRON_NODE = 'CJ_CRON_NODE',
}

export enum TingEnum {
  REGISTER_STATUS = 'TING_REGISTER_STATUS',
  CONTAINER_STATUS = 'TING_CONTAINER_STATUS',
  CURRENT_STATUS = 'TING_CURRENT_STATUS',
}

export enum EcdnEnum {
  SERVER_ROOM_TYPE = 'ECDN_SERVER_ROOM_TYPE',
  SERVER_STATUS = 'ECDN_SERVER_STATUS',
}

export enum ZEnum {
  Z_ISP_NAME = 'Z_ISP_NAME',
  Z_DEVICE_STATUS = 'Z_DEVICE_STATUS',
  Z_DEVICE_RUN_STATES = 'Z_DEVICE_RUN_STATES',
  Z_DELIVERY_STATUS = 'Z_DELIVERY_STATUS',
  Z_SUBMIT_TYPE = 'Z_SUBMIT_TYPE',
  Z_DEVICE_LOGIN_TYPE = 'Z_DEVICE_LOGIN_TYPE',
  Z_DELIVERY_RESOURCE_TYPE = 'Z_DELIVERY_RESOURCE_TYPE',
  Z_MACHINE_TYPE = 'Z_MACHINE_TYPE',
  Z_ACCOUNT_TYPE = 'Z_ACCOUNT_TYPE',

  ZP_ACCEPTANCE_STATUS = 'ZP_ACCEPTANCE_STATUS',
  ZP_DELIVERY_STATUS = 'ZP_DELIVERY_STATUS',
  ZP_DELIVERY_TYPE = 'ZP_DELIVERY_TYPE',
}

export enum SAEnum {
  SA_RESOURCE_GAP_STATUS = 'SA_RESOURCE_GAP_STATUS',
  SA_UTILIZATION_VERSION = 'SA_UTILIZATION_VERSION',
}

export enum BEnum {
  DEVICE_TYPE = 'B_DEVICE_TYPE',
  IP_TYPE = 'B_IP_TYPE',
  AGENT_STATUS = 'B_AGENT_STATUS',
  BUSINESS_STATUS = 'B_BUSINESS_STATUS',
  ERROR_TYPE = 'B_ERROR_TYPE',
  LOCAL_COVERAGE_ONLY = 'B_LOCAL_COVERAGE_ONLY',
}

export enum LBEnum {
  BILL_TYPE = 'LB_BILL_TYPE',
  DEVICE_STATUS = 'LB_DEVICE_STATUS',
}

export enum CommonEnum {
  ISP = 'COMMON_ISP',
  YES_NO = 'COMMON_YES_NO',
  MERGE_ISP = 'COMMON_MERGE_ISP',
  ONLINE_STATUS = 'COMMON_ONLINE_STATUS',
}

export enum InspectEnum {
  OPERATOR_TYPE = 'INSPECT_OPERATOR_TYPE',
  EXEC_STATUS = 'INSPECT_EXEC_STATUS',
  INSPECT_RESULT = 'INSPECT_INSPECT_RESULT',
  IDC_TYPE = 'INSPECT_IDC_TYPE',
}

export enum LAEnum {
  IS_DIFF_ISP = 'LA_IS_DIFF_ISP',
  BUSINESS_STATUS = 'LA_BUSINESS_STATUS',
}

export enum RoomEnum {
  STATUS = 'ROOM_STATUS',
  CHARGE_MODE = 'ROOM_CHARGE_MODE',
  PEAK_ENABLE_STATUS = 'ROOM_PEAK_ENABLE_STATUS',
  TYPE = 'ROOM_TYPE',
  USE_MODE = 'ROOM_USE_MODE',
  SPEED_LIMIT_STATUS = 'ROOM_SPEED_LIMIT_STATUS',
  SPEED_LIMIT_OPER_TYPE = 'ROOM_SPEED_LIMIT_OPER_TYPE',
  SPEED_LIMIT_OPER_STATUS = 'ROOM_SPEED_LIMIT_OPER_STATUS',
  NETWORK_SPEED_LIMIT_JOB_TYPE = 'NETWORK_SPEED_LIMIT_JOB_TYPE',
  NETWORK_SPEED_LIMIT_JOB_STATUS = 'NETWORK_SPEED_LIMIT_JOB_STATUS',
  NETWORK_SPEED_LIMIT_JOB_EXEC_STATUS = 'NETWORK_SPEED_LIMIT_JOB_EXEC_STATUS',
  NETWORK_SPEED_LIMIT_TARGET_TYPE = 'NETWORK_SPEED_LIMIT_TARGET_TYPE',
  NETWORK_SPEED_LIMIT_LIMIT_TYPE = 'NETWORK_SPEED_LIMIT_LIMIT_TYPE',
}

export enum ZCDNEnum {
  ZCDN_BIZTYPE = 'ZCDN_BIZTYPE',
}